# openapi.yaml
openapi: 3.0.0

info:
    title: Pipes PHP SDK
    description: Pipes PHP SDK API
    version: '1.0'

servers:
    -   url: http://127.0.0.4
        description: Docker Localhost

paths:
    /applications:
        get:
            tags: [ Applications ]
            responses:
                "200":
                    $ref: '#/components/responses/Empty'

    /applications/sdk/{sdk}/limits:
        post:
            tags: [ Applications]
            parameters:
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "200":
                    $ref: '#/components/responses/Limits'
    /applications/{key}:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
    /applications/{key}/sync/list:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
    /applications/{key}/sync/{method}:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/Method'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
        post:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/Method'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
    /applications/{key}/users/{user}/sdk/{sdk}/authorize:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}/authorize/token:
        post:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "302":
                    description: ''
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/authorize/token:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/State'
            responses:
                "302":
                    description: ''
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/users/{user}/sdk/{sdk}:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstalls'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}:
        get:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstall'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}/install:
        post:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            requestBody:
                $ref: '#/components/requestBodies/Empty'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstallInstall'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}/settings:
        put:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            requestBody:
                $ref: '#/components/requestBodies/Settings'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstallSettings'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}/password:
        put:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            requestBody:
                $ref: '#/components/requestBodies/Password'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstallUninstall'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /applications/{key}/users/{user}/sdk/{sdk}/uninstall:
        delete:
            tags: [ Applications ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            responses:
                "200":
                    $ref: '#/components/responses/ApplicationInstallUninstall'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /connector/list:
        get:
            tags: [ Connectors ]
            responses:
                "200":
                    $ref: '#/components/responses/Connectors'
                "500":
                    $ref: '#/components/responses/500'

    /connector/{id}/action:
        post:
            tags: [ Connectors ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /connector/{id}/action/test:
        get:
            tags: [ Connectors ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /connector/{id}/webhook:
        post:
            tags: [ Connectors ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /connector/{id}/webhook/test:
        get:
            tags: [ Connectors ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /custom-node/list:
        get:
            tags: [ Custom Nodes ]
            responses:
                "200":
                    $ref: '#/components/responses/Connectors'
                "500":
                    $ref: '#/components/responses/500'

    /custom-node/{id}/process:
        post:
            tags: [ Custom Nodes ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /custom-node/{id}/process/test:
        get:
            tags: [ Custom Nodes ]
            parameters:
                -   $ref: '#/components/parameters/Id'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'



    /webhook/applications/{key}/users/{user}/sdk/{sdk}/subscribe:
        post:
            tags: [ Webhooks ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            requestBody:
                $ref: '#/components/requestBodies/Webhook'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /webhook/applications/{key}/users/{user}/sdk/{sdk}/unsubscribe:
        post:
            tags: [ Webhooks ]
            parameters:
                -   $ref: '#/components/parameters/Key'
                -   $ref: '#/components/parameters/User'
                -   $ref: '#/components/parameters/Sdk'
            requestBody:
                $ref: '#/components/requestBodies/Webhook'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "404":
                    $ref: '#/components/responses/404'
                "500":
                    $ref: '#/components/responses/500'

    /parser/{type}/to/json:
        post:
            tags: [ Parsers ]
            parameters:
                -   $ref: '#/components/parameters/Type'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /parser/{type}/to/json/test:
        get:
            tags: [ Parsers ]
            parameters:
                -   $ref: '#/components/parameters/Type'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /parser/json/to/{type}:
        post:
            tags: [ Parsers ]
            parameters:
                -   $ref: '#/components/parameters/Type'
            responses:
                "200":
                    $ref: '#/components/responses/Empty'
                "500":
                    $ref: '#/components/responses/500'

    /parser/json/to/{type}/test:
        get:
            tags: [ Parsers ]
            parameters:
                -   $ref: '#/components/parameters/Type'
            responses:

components:
    responses:
        Application:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Application'

        Applications:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Applications'

        ApplicationInstall:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApplicationInstall'

        ApplicationInstalls:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApplicationInstalls'

        ApplicationInstallInstall:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApplicationInstall'

        ApplicationInstallUninstall:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApplicationInstallUninstall'

        ApplicationInstallSettings:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ApplicationInstallSettings'

        Empty:
            description: ''
            content:
                application/json:
                    schema:
                        type: object

        Limits:
            description: ''
            content:
                application/json:
                    schema:
                        type: array
                        items:
                            type: string
                            example: "user|system;10;60"
        Connectors:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Connectors'

        TopologyTasks:
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/TopologyTasks'

        "404":
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Message'
                    example:
                        type: AppStoreException
                        status: NOT_FOUND
                        message: Application with key 'Unknown' not found!
                        error_code: 404

        "500":
            description: ''
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Message'
                    example:
                        type: AppStoreException
                        status: INTERNAL_SERVER_ERROR
                        message: Something gone wrong
                        error_code: 500

    schemas:
        Application:
            type: object
            properties:
                key:
                    type: string
                name:
                    type: string
                description:
                    type: string
                application_type:
                    type: string
                    enum: [cron, webhook]
                authorization_type:
                    type: string
                    enum: [basic, oauth, oauth2]

        Applications:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Application'

        ApplicationInstall:
            type: object
            properties:
                key:
                    type: string
                name:
                    type: string
                description:
                    type: string
                application_type:
                    type: string
                    enum: [cron, webhook]
                authorization_type:
                    type: string
                    enum: [basic, oauth, oauth2]
                authorized:
                    type: boolean
                applicationSettings:
                    type: array
                    items:
                        $ref: '#/components/schemas/ApplicationSetting'
                webhookSettings:
                    type: array
                    items:
                        $ref: '#/components/schemas/WebhookSetting'

        ApplicationInstallUninstall:
            type: object
            properties:
                id:
                    type: string
                key:
                    type: string
                user:
                    type: string
                authorized:
                    type: boolean
                settings:
                    type: object
                nonEncryptedSettings:
                    type: object
                created:
                    type: string
                updated:
                    type: string
                expires:
                    type: string
                    nullable: true

        ApplicationInstalls:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/ApplicationInstallUninstall'

        ApplicationInstallSettings:
            type: object
            properties:
                id:
                    type: string
                key:
                    type: string
                user:
                    type: string
                settings:
                    type: object
                nonEncryptedSettings:
                    type: object
                created:
                    type: string
                updated:
                    type: string
                expires:
                    type: string
                    nullable: true
                applicationSettings:
                    type: array
                    items:
                        $ref: '#/components/schemas/ApplicationSetting'

        ApplicationSetting:
            type: object
            properties:
                key:
                    type: string
                type:
                    type: string
                    enum: [text, number, url, password, selectbox, checkbox]
                label:
                    type: string
                value:
                    type: string
                description:
                    type: string
                required:
                    type: boolean
                readOnly:
                    type: boolean
                disabled:
                    type: boolean
                choices:
                    type: array
                    items:
                        type: string

        Connectors:
            type: array
            items:
                type: object
                properties:
                    name:
                        type: string
                        example: 'get-user-connector'
                    app:
                        type: string
                        example: 'shoptet'

        TopologyTask:
            type: object
            properties:
                created:
                    type: string
                updated:
                    type: string
                data:
                    type: string
                    format: json
                auditLogs:
                    type: string
                    format: json
                topologyId:
                    type: string
                topologyName:
                    type: string
                nodeId:
                    type: string
                nodeName:
                    type: string
                processId:
                    type: string

        TopologyTasks:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/TopologyTask'
                count:
                    type: integer
                total:
                    type: integer
                limit:
                    type: integer
                offset:
                    type: integer

        Message:
            type: object
            properties:
                type:
                    type: string
                status:
                    type: string
                message:
                    type: string
                error_code:
                    type: integer

        WebhookSetting:
            type: object
            properties:
                name:
                    type: string
                topology:
                    type: string
                default:
                    type: boolean
                enabled:
                    type: boolean

    parameters:
        Id:
            name: id
            in: path
            required: true
            example: id
            schema:
                type: string

        Key:
            name: key
            in: path
            required: true
            example: id
            schema:
                type: string

        User:
            name: user
            in: path
            required: true
            example: user
            schema:
                type: string

        Topology:
            name: topology
            in: path
            required: true
            example: topology
            schema:
                type: string

        Node:
            name: node
            in: path
            required: true
            example: node
            schema:
                type: string

        Type:
            name: type
            in: path
            required: true
            example: type
            schema:
                type: string
                enum: [xlsx, xls, ods, csv, html]

        State:
            name: state
            in: query
            required: true
            example: state
            schema:
                type: string

        Sdk:
            name: sdk
            in: path
            required: true
            example: sdk
            schema:
                type: string

        Method:
            name: method
            in: path
            required: true
            example: method
            schema:
                type: string